requirements_linux_dev.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cycler==0.11.0
deepface==0.0.75
fastapi==0.88.0
filelock==3.8.2
Flask==2.2.2
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphql-core==3.2.3
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mtcnn==0.1.1
neo4j==5.3.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
scipy==1.9.3
seaborn==0.11.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
strawberry-graphql==0.149.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
